MISUV 4026: moved tax year methods to utils
